refactor: Comp -> Login 폴더명 변경
